First Commit
This is an APK of a flutter app where users can enter the name of their favorite actor along with the actor's photo from the device's gallery or a new one captured from the camera. Upon saving the data it displays the picture and name of the user's favorite actor on a new page.